Move pipelines responses to package folder
