+++ src/main/java/io/seqera/tower/cli/responses/pipelines/PipelinesList.java
@@ -1,5 +1,6 @@
-package io.seqera.tower.cli.responses;
+package io.seqera.tower.cli.responses.pipelines;
 
+import io.seqera.tower.cli.responses.Response;
 import io.seqera.tower.cli.utils.TableList;
 import io.seqera.tower.model.PipelineDbDto;
