+++ src/main/java/io/seqera/tower/cli/responses/pipelines/PipelinesUpdated.java
@@ -1,4 +1,6 @@
-package io.seqera.tower.cli.responses;
+package io.seqera.tower.cli.responses.pipelines;
+
+import io.seqera.tower.cli.responses.Response;
 
 public class PipelinesUpdated extends Response {
