+++ src/main/java/io/seqera/tower/cli/responses/pipelines/PipelinesView.java
@@ -1,7 +1,8 @@
-package io.seqera.tower.cli.responses;
+package io.seqera.tower.cli.responses.pipelines;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import io.seqera.tower.JSON;
+import io.seqera.tower.cli.responses.Response;
 import io.seqera.tower.cli.utils.ModelHelper;
 import io.seqera.tower.cli.utils.TableList;
 import io.seqera.tower.model.Launch;
