+++ src/test/java/io/seqera/tower/cli/PipelinesCmdTest.java
@@ -9,11 +9,11 @@ import io.seqera.tower.cli.exceptions.NoComputeEnvironmentException;
 import io.seqera.tower.cli.exceptions.PipelineNotFoundException;
 import io.seqera.tower.cli.exceptions.TowerException;
 import io.seqera.tower.cli.exceptions.WorkspaceNotFoundException;
-import io.seqera.tower.cli.responses.PipelinesCreated;
-import io.seqera.tower.cli.responses.PipelinesDeleted;
-import io.seqera.tower.cli.responses.PipelinesList;
-import io.seqera.tower.cli.responses.PipelinesUpdated;
-import io.seqera.tower.cli.responses.PipelinesView;
+import io.seqera.tower.cli.responses.pipelines.PipelinesCreated;
+import io.seqera.tower.cli.responses.pipelines.PipelinesDeleted;
+import io.seqera.tower.cli.responses.pipelines.PipelinesList;
+import io.seqera.tower.cli.responses.pipelines.PipelinesUpdated;
+import io.seqera.tower.cli.responses.pipelines.PipelinesView;
 import io.seqera.tower.model.ComputeEnv;
 import io.seqera.tower.model.Launch;
 import io.seqera.tower.model.PipelineDbDto;
